Being fleshing out support for alternate configuration files and a configuration file template.
This is very hackish and custom for now - it implements just
ethernet and hostname overriding.  But this is needed for
at least the carambola2 as the development board doesn't wire arge0 up to
an external PHY - only arge1 is hooked up to the ethernet switch
internally, and switch ports 1/2 are wired up to the actual
RJ45 jacks.