Fix netboot target
The netboot target is broken since mfs image embedding no longer
works the way it's currently written.

This commit fixes the netboot target for cases that use the old
style mfs image embedding (by specifying the space to reserve
for mfs within the kernel via MD_ROOT_SIZE in kernel configs).

We should really look at moving to the new style of mfs image
embedding that does all this dynamically and uses only the required
amount of space within the kernel for mfs, but that's for another
time.