Initial hacks to experiment with cross building packages and hten putting the results in a format we can hopefully install into the rootfs.
It doesn't /entirely/ work yet because there's no way to populate
the manifest file in a portable way.  It turns out that tar
populates and reads mtree files differently to what makefs wants,
and this is a bit of a problem.  Sigh.

So, this is a WIP, and I need to finish hacking on it a bunch
more.